Restructuring to move abacus to modules
