+++ wandering-warriors/modules/__init__.py
@@ -1,4 +1,5 @@
 from .draw_pad import DrawPad
 from .ledger import LedgerLayout
+from .abacus import Abacus
 
-__all__ = ('DrawPad', 'LedgerLayout')
+__all__ = ('DrawPad', 'LedgerLayout', 'Abacus')
